requirements.txt
Pillow==12.2.0
python-dotenv==1.2.2